Extract the one GlobalState publishing protocol
SonarCloud measured 9.1% duplicated lines on the new code in #569 against
a 0.4% project baseline, and it named the blocks: catalogue.py:67 and
:183, _data_frame_source.py:91 and :161, plus the cross-file pairs.  Four
copies of one protocol, because I wrote it once and pasted it three times.

They are character-for-character identical apart from a comment, and the
protocol is fiddly in exactly the corners that make copies dangerous:
restoring a shadowed entry, closing only a scope this block opened, and
holding both on the error path.  Four copies is four places for those to
drift.

_published_in_global_state carries it once, beside the scope guard it
uses.  A subclass supplies _STATE_KEY and may override _publish /
_unpublish when it has more to do than store itself there -- which is
what would let the persistence storage join later without a fifth copy.

The extracted version also closes the scope if publishing raises, which
none of the four copies did.
