+++ python/hgraph/adaptors/data_frame/_data_frame_source.py
@@ -6,7 +6,8 @@ from typing import Iterator, TypeVar
 import pyarrow as pa
 
 from hgraph import GlobalState
-from hgraph._wiring._state import _active_global_state, _global_state_scope
+from hgraph._wiring._state import (_active_global_state, _global_state_scope,
+                                   _published_in_global_state)
 
 __all__ = (
     "DataFrameSource",
@@ -54,7 +55,7 @@ class DataFrameSource(ABC):
 DATA_FRAME_SOURCE = TypeVar("DATA_FRAME_SOURCE", bound=DataFrameSource)
 
 
-class DataStore:
+class DataStore(_published_in_global_state):
     """Data-source instances held in the Python seed/result GlobalState."""
 
     _STATE_KEY = ":adaptors:data_frame:sources"
@@ -88,29 +89,6 @@ class DataStore:
             self._data_frame_sources[dfs] = source
         return source
 
-    def __enter__(self):
-        # This block publishes the store into the GlobalState for the wiring
-        # inside it to find, so it needs a state to publish into.  Open one for
-        # the block when the caller has not, and close it in __exit__ -- the
-        # state must not outlive the `with` that needed it.
-        self._state_scope = _global_state_scope()
-        state = self._state_scope.__enter__()
-        self._previous = state.get(self._STATE_KEY, self._MISSING)
-        state[self._STATE_KEY] = self
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        state = _active_global_state()
-        if self._previous is self._MISSING:
-            state.pop(self._STATE_KEY, None)
-        else:
-            state[self._STATE_KEY] = self._previous
-        self._previous = self._MISSING
-        scope, self._state_scope = self._state_scope, None
-        if scope is not None:
-            scope.__exit__(exc_type, exc_value, traceback)
-        return False
-
 
 class ArrowDataFrameSource(DataFrameSource):
     def __init__(self, frame):
@@ -129,7 +107,7 @@ class PolarsDataFrameSource(ArrowDataFrameSource):
         super().__init__(df)
 
 
-class DataConnectionStore:
+class DataConnectionStore(_published_in_global_state):
     _STATE_KEY = ":adaptors:data_frame:connections"
     _MISSING = object()
 
@@ -157,29 +135,6 @@ class DataConnectionStore:
 
     def set_connection(self, name: str, connection):
         self._connections[name] = connection
-
-    def __enter__(self):
-        # This block publishes the store into the GlobalState for the wiring
-        # inside it to find, so it needs a state to publish into.  Open one for
-        # the block when the caller has not, and close it in __exit__ -- the
-        # state must not outlive the `with` that needed it.
-        self._state_scope = _global_state_scope()
-        state = self._state_scope.__enter__()
-        self._previous = state.get(self._STATE_KEY, self._MISSING)
-        state[self._STATE_KEY] = self
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        state = _active_global_state()
-        if self._previous is self._MISSING:
-            state.pop(self._STATE_KEY, None)
-        else:
-            state[self._STATE_KEY] = self._previous
-        self._previous = self._MISSING
-        scope, self._state_scope = self._state_scope, None
-        if scope is not None:
-            scope.__exit__(exc_type, exc_value, traceback)
-        return False
 
 
 class SqlDataFrameSource(DataFrameSource):
